EhCache

+++ autorest-rev/src/main/java/me/mindex/autorest/persistence/repo/YtmWordListRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 
-public interface YtmWordListRepo extends JpaRepository<YtmWordListEntity, Integer>, JpaSpecificationExecutor<YtmWordListEntity> {
+public interface YtmWordListRepo extends JpaRepository<YtmWordListEntity, Long>, JpaSpecificationExecutor<YtmWordListEntity> {
 
 	YtmWordListEntity findByWord(String word);
 }
